WIP on revisions for AST
