+++ package.json
@@ -5,15 +5,18 @@
   "main": "./src/main/main.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack serve --hot & tailwindcss -i ./src/renderer/App.css -o ./src/renderer/dist/output.css --watch & npx electronmon --trace-warning .",
+    "start": "webpack serve --hot & npx electronmon --trace-warning .",
     "electron": "electron .",
     "proxy": "nodemon proxy/proxy.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@babel/parser": "^7.21.3",
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-react": "^7.18.6",
+    "@babel/traverse": "^7.21.3",
+    "@babel/types": "^7.21.3",
     "babel-loader": "^9.1.2",
     "css-loader": "^6.7.3",
     "daisyui": "^2.51.3",
