+++ webpack.config.js
@@ -9,6 +9,7 @@ module.exports = {
   resolve: {
     extensions: [
       '.js',
+      '.ts',
       '.tsx',
       '.png',
       '.jpg',
